requirements.txt
# Flask
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# MongoDB
pymongo==4.6.3

# Password hashing
bcrypt==4.1.2

# AI (optional)
openai

# Utilities
python-dateutil==2.8.2

# Production server
gunicorn==21.2.0